+++ src/main/java/com/library/repository/LigneAvoirRepository.java
@@ -1,9 +1,6 @@
 package com.library.repository;
 
 import com.library.entities.LigneAvoir;
-import com.library.entities.LigneCmdClient;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -15,11 +12,11 @@ import java.util.List;
 @Repository
 public interface LigneAvoirRepository extends JpaRepository<LigneAvoir, Long> {
 
-    List<LigneAvoir> findAllByNumero(int numero);
+    List<LigneAvoir> findAllByNumero(long numero);
 
     @Modifying
     @Query("delete from LigneAvoir where numero = :numero")
-    void deleteByNumero(@Param("numero") int numero);
+    void deleteByNumero(@Param("numero") long numero);
 
     @Query("select p from LigneAvoir p where p.produit.id =:prod")
     List<LigneAvoir> ListLigneAvoirByProduitId(@Param("prod") Long prodId);
