+++ src/main/java/com/library/repository/LigneDevisRepository.java
@@ -1,9 +1,6 @@
 package com.library.repository;
 
-import com.library.entities.LigneCmdClient;
 import com.library.entities.LigneDevis;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -15,11 +12,11 @@ import java.util.List;
 @Repository
 public interface LigneDevisRepository extends JpaRepository<LigneDevis, Long> {
 
-    List<LigneDevis> findAllByNumero(int numero);
+    List<LigneDevis> findAllByNumero(long numero);
 
     @Modifying
     @Query("delete from LigneDevis where numero = :numero")
-    void deleteByNumero(@Param("numero") int numero);
+    void deleteByNumero(@Param("numero") long numero);
 
     @Query("select p from LigneDevis p where p.produit.id =:prod")
     List<LigneDevis> ListLigneDevisByProduitId(@Param("prod") Long prodId);
