+++ src/main/java/com/library/services/impl/LigneAvoirServiceImpl.java
@@ -1,16 +1,10 @@
 package com.library.services.impl;
 
 import com.library.entities.LigneAvoir;
-import com.library.entities.LigneCmdClient;
 import com.library.exceptions.ResourceNotFoundException;
 import com.library.repository.LigneAvoirRepository;
-import com.library.repository.LigneCmdClientRepository;
 import com.library.services.LigneAvoirService;
-import com.library.services.LigneCmdClientService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -64,7 +58,7 @@ public class LigneAvoirServiceImpl implements LigneAvoirService {
     }
 
     @Override
-    public void deleteLavoirByNumero(int numero) {
+    public void deleteLavoirByNumero(long numero) {
         ligneAvoirRepository.deleteByNumero(numero);
     }
 
@@ -79,7 +73,7 @@ public class LigneAvoirServiceImpl implements LigneAvoirService {
     }
 
     @Override
-    public List<LigneAvoir> findAllLavoirByNumero(int numero) {
+    public List<LigneAvoir> findAllLavoirByNumero(long numero) {
         return ligneAvoirRepository.findAllByNumero(numero);
     }
 }
