+++ src/main/java/com/library/services/impl/LigneCmdClientServiceImpl.java
@@ -1,21 +1,18 @@
 package com.library.services.impl;
 
-import java.util.List;
-import java.util.Optional;
-
-
-import com.library.entities.Produit;
+import com.library.entities.LigneCmdClient;
+import com.library.exceptions.ResourceNotFoundException;
+import com.library.repository.LigneCmdClientRepository;
 import com.library.services.LigneCmdClientService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-
-import com.library.entities.LigneCmdClient;
-import com.library.exceptions.ResourceNotFoundException;
-import com.library.repository.LigneCmdClientRepository;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -74,7 +71,7 @@ public class LigneCmdClientServiceImpl implements LigneCmdClientService {
     }
 
     @Override
-    public void deleteLcomByNumero(int numero) {
+    public void deleteLcomByNumero(long numero) {
         ligneCmdClientRepository.deleteByNumero(numero);
     }
 
@@ -89,7 +86,7 @@ public class LigneCmdClientServiceImpl implements LigneCmdClientService {
     }
 
     @Override
-    public List<LigneCmdClient> findAllLcomByNumero(int numero) {
+    public List<LigneCmdClient> findAllLcomByNumero(long numero) {
         return ligneCmdClientRepository.findAllByNumero(numero);
     }
