+++ src/main/java/com/library/services/impl/LigneCreanceServiceImpl.java
@@ -1,15 +1,10 @@
 package com.library.services.impl;
 
-import com.library.entities.LigneCmdClient;
 import com.library.entities.LigneCreance;
 import com.library.exceptions.ResourceNotFoundException;
-import com.library.repository.LigneCmdClientRepository;
 import com.library.repository.LigneCreanceRepository;
-import com.library.services.LigneCmdClientService;
 import com.library.services.LigneCreanceService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -69,7 +64,7 @@ public class LigneCreanceServiceImpl implements LigneCreanceService {
     }
 
     @Override
-    public void deleteLcreanceByNumero(int numero) {
+    public void deleteLcreanceByNumero(long numero) {
         ligneCreanceRepository.deleteByNumero(numero);
     }
 
@@ -84,7 +79,7 @@ public class LigneCreanceServiceImpl implements LigneCreanceService {
     }
 
     @Override
-    public List<LigneCreance> findAllLcreanceByNumero(int numero) {
+    public List<LigneCreance> findAllLcreanceByNumero(long numero) {
         return ligneCreanceRepository.findAllByNumero(numero);
     }
 }
